requirements.txt
asgiref==3.4.1
Django==3.2.5
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
djangorestframework~=3.12.4
